Update workflows
- Don't generate release notes unless adding a draft
- Generate man page in Arch Linux PKGBUILD
- Update build script so man pages refer to `pretty-php`, not
  `pretty-php.phar`, and remove unnecessary `git clone` step if running
  in a CI environment

+++ scripts/build.sh
@@ -17,8 +17,9 @@ function usage() {
     else
         cat
     fi <<EOF
-usage: ${0##*/}                   build the current working tree
+usage: ${0##*/} [worktree]        build the current working tree
        ${0##*/} man               build a man page for the current working tree
+       ${0##*/} man worktree      build the working tree and a man page for it
        ${0##*/} [man] latest      build the closest version reachable from HEAD
        ${0##*/} [man] v<VERSION>  build a tagged version${1:+
 }
@@ -44,6 +45,11 @@ FROM_GIT=0
 VERSION=
 while (($#)); do
     case "$1" in
+    worktree)
+        BUILD_PHAR=1
+        FROM_GIT=0
+        VERSION=
+        ;;
     latest)
         BUILD_PHAR=1
         FROM_GIT=1
@@ -67,9 +73,15 @@ while (($#)); do
     shift
 done
 
-TEMP_DIR=$(mktemp -d) &&
-    trap 'rm -rf "$TEMP_DIR"' EXIT ||
-    die "error creating temporary directory"
+if [[ ${CI-} == true ]] && ((!FROM_GIT)); then
+    printf '==> building in CI environment: %s\n' "$REPO"
+    TEMP_DIR=$REPO
+    rm -rf "$DIST"
+else
+    TEMP_DIR=$(mktemp -d) &&
+        trap 'rm -rf "$TEMP_DIR"' EXIT ||
+        die "error creating temporary directory"
+fi
 
 if ((FROM_GIT)); then
     # Get the closest annotated version tag reachable from HEAD
@@ -86,7 +98,7 @@ if ((FROM_GIT)); then
         git -C "$TEMP_DIR" fetch --progress --no-recurse-submodules --depth=1 origin +refs/tags/"$VERSION":refs/tags/"$VERSION" &&
         git -C "$TEMP_DIR" -c advice.detachedHead=false checkout --progress --force refs/tags/"$VERSION" ||
         die "error checking out $PACKAGE $VERSION"
-elif ((BUILD_PHAR)); then
+elif ((BUILD_PHAR)) && [[ $TEMP_DIR != "$REPO" ]]; then
     printf '==> copying %s working tree to %s\n' "$PACKAGE" "$TEMP_DIR"
     cp -Lpr "$REPO"/!(box|build|vendor) "$TEMP_DIR" ||
         die "error copying working tree to $TEMP_DIR"
@@ -106,22 +118,37 @@ if ((BUILD_PHAR)); then
         php -d phar.readonly=off box/vendor/bin/box compile -d "$TEMP_DIR" --no-interaction
 
     printf '==> finalising build\n'
-    TEMP_PHAR=("$TEMP_DIR/$DIST"/*)
+    TEMP_PHAR=("$TEMP_DIR/$DIST"/*.phar)
     [[ ${#TEMP_PHAR[@]} -eq 1 ]] ||
         die "output missing or invalid"
-    rm -f "$PHAR" &&
-        mkdir -pv "$DIST" &&
-        cp -pv "$TEMP_PHAR" "$PHAR" ||
-        die "error copying $TEMP_PHAR to $PHAR"
+    if [[ ! $PHAR -ef $TEMP_PHAR ]]; then
+        rm -f "$PHAR" &&
+            mkdir -pv "$DIST" &&
+            cp -pv "$TEMP_PHAR" "$PHAR" ||
+            die "error copying $TEMP_PHAR to $PHAR"
+    fi
 
     printf ' -> PHP archive created at %s/%s\n' "$REPO" "$PHAR"
 fi
 
 if ((BUILD_MAN)); then
-    ((BUILD_PHAR)) || PHAR=bin/pretty-php
-    printf '==> generating man page for %s\n' "$PHAR"
+    # Run the script from the bin directory to ensure the command name in the
+    # man page isn't <package>.phar
+    if ((BUILD_PHAR)); then
+        RUN=$TEMP_DIR/bin/$PACKAGE
+    else
+        RUN=bin/$PACKAGE
+        # If packages are installed, assume they're current, otherwise install
+        # production dependencies
+        if ! composer show 2>/dev/null | grep . >/dev/null; then
+            printf '==> installing %s production dependencies in %s\n' "$PACKAGE" "$REPO"
+            composer install --no-plugins --no-interaction --no-dev
+        fi
+    fi
+    printf '==> generating man page for %s\n' "$RUN"
     rm -f "$MAN" &&
-        "$PHAR" _man | pandoc --standalone --to man -o "$MAN" ||
+        mkdir -pv "$DIST" &&
+        "$RUN" _man | pandoc --standalone --to man -o "$MAN" ||
         die "error creating man page at $MAN"
 
     printf ' -> man page created at %s/%s\n' "$REPO" "$MAN"
